Add parameter removedJobAction to the createJobs method
The default value is DELETE that was the value before the change.
This make it possible to define this action during the Jobs run
according to the needs.

One possible scenario if just to add new jobs and keep the old once
that can be achieved with 'IGNORE'.

Or keep the old once but disable them for backup reasons. Can be done
with 'DISABLE'.